jumps: carry a 32-bit displacement — past 32KB of body every loop and branch jumped somewhere arbitrary
`OpGotoWord` / `OpGotoFalseWord` carried a `const i16` displacement, written with
an unchecked `as i16` at every emission site.  Past ~32 KB of emitted body the
value wrapped, so a `while true` took its backward jump to an arbitrary address:
the body ran ONCE, control fell out of the loop, and main returned 0 with no
diagnostic.  Every println inside still ran exactly once, so a log ended
mid-story with nothing wrong in it.

The filed scope was the backward `while` jump.  A boundary matrix found the real
one: ALL FIVE jump classes truncate, because they share the encoding — `while`
and `for` (backward), `break` (forward, patched in Stack::end_loop), and the
forward skips of `if` and `else`.  `--native` was correct in all five cells, since
it emits real Rust control flow and never reads these operands; that made it the
positive control the matrix was read against.

Both ops now carry `const i32`, which covers the whole `code_pos` (u32) space —
the threshold is removed rather than moved to a larger number.  A fixed-width slot
also means the forward-jump patches need no branch relaxation: code_put writes an
i32 into a slot already reserved at that size.

Emitting a 4-byte constant needed two places to stop deriving an integer's width
from its RANGE and read its declared `size(N)` instead: `variables::size` in the
Constant context, and `Data::rust_type`, which decides what the generated reader
in fill.rs reads — left alone it would have READ an i64 while codegen WROTE 4
bytes, the reader and writer disagreeing about one operand.  Both are inert for
every alias that predates them (u8/i8 force 1 and range to 1, u16/i16 force 2 and
range to 2, plain `integer` forces nothing), proven by a byte-identical
`loft introspect` over a corpus of all of them before and after that change alone.

Displacements measured from after-the-operand move from -2 to -4; the
disassembler's target scan and both debug renderers follow the same width.
tests/dumps is gitignored, so no golden output churned.

Guard: one case per jump class, each asserting an accumulated VALUE rather than
mere completion — the failure mode was silent fall-through, which a
runs-to-completion check would have called a pass.  Verified non-vacuous against
the installed pre-change 2026.7.2 binary: all five produce NO output there.

Suite 3528/3528.  Verified at 10x the old threshold (~350 KB of body).

Fixes #654

+++ src/fill.rs
@@ -318,8 +318,8 @@ fn goto(s: &mut State) {
 }
 
 fn goto_word(s: &mut State) {
-    let v_step = s.code::<i16>();
-    s.code_pos = (s.code_pos as i32 + i32::from(v_step)) as u32;
+    let v_step = s.code::<i32>();
+    s.code_pos = (i64::from(s.code_pos) + i64::from(v_step)) as u32;
 }
 
 fn goto_false(s: &mut State) {
@@ -331,10 +331,10 @@ fn goto_false(s: &mut State) {
 }
 
 fn goto_false_word(s: &mut State) {
-    let v_step = s.code::<i16>();
+    let v_step = s.code::<i32>();
     let v_if_false = *s.get_stack::<u8>();
     if v_if_false != 1 {
-        s.code_pos = (s.code_pos as i32 + i32::from(v_step)) as u32;
+        s.code_pos = (i64::from(s.code_pos) + i64::from(v_step)) as u32;
     }
 }
 
@@ -1252,7 +1252,7 @@ fn text_compare(s: &mut State) {
 }
 
 fn cast_character_from_int(s: &mut State) {
-    let v_v1 = *s.get_stack::<i64>();
+    let v_v1 = *s.get_stack::<i32>();
     let new_value = if let Some(c) = char::from_u32(v_v1 as u32) {
         c
     } else {
